% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{get_reference}
\alias{get_reference}
\title{Get reference data frame}
\usage{
get_reference(antigen, epitope, org, binding = TRUE)
}
\arguments{
\item{antigen}{The antigen to get the reference data frame for}

\item{epitope}{The epitope to get the reference data frame for}

\item{org}{The organism to get the reference data frame for}

\item{binding}{Whether the BCR is binding to the antigen}
}
\value{
A data frame with the reference data
}
\description{
Get reference data frame
}
